Rename: connect.py => clockify.py
